+++ actions/index.ts
@@ -43,7 +43,6 @@ export { getClinicSchedule } from "./clinic-schedule";
 export { createMedication, getMedicationsList } from "./medications";
 export { sendEmail, sendSimpleEmail } from "./send-email";
 export { getEducationalContent, getEducationalContentByUuid } from "./educational-content";
-export { getUserNotifications, getUserNotification, markNotificationAsRead } from "./notification";
 export { addPrescription, viewPrescription, deletePrescription } from "./prescription";
 export {
     changeTheme,
@@ -65,3 +64,11 @@ export {
     getUpcomingPrescriptions,
     getUpcomingVaccinations,
 } from "./dashboard-healthcare";
+export {
+    createNotification,
+    deleteNotification,
+    getUserNotification,
+    getUserNotifications,
+    markAllNotificationsAsRead,
+    markNotificationAsRead,
+} from "./notification";
