+++ actions/veterinary.ts
@@ -1,11 +1,10 @@
 "use server";
 import { VeterinarianSchema, VeterinarianType } from "@/schemas";
-import { generateVerificationToken, hashPassword } from "@/lib";
-import { prisma } from "@/lib";
+import { generateVerificationToken, hashPassword, prisma } from "@/lib";
 import { createNewPreferenceDefault } from "@/actions";
 import { role_type, type users, type veterinarians } from "@prisma/client";
 import { type veterinary_specialization } from "@prisma/client";
-import type { ActionResponse } from "@/types/server-action-response";
+import type { ActionResponse } from "@/types";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { redirect } from "next/navigation";
