+++ lib/index.ts
@@ -11,3 +11,5 @@ export { toTitleCase } from "./functions/text/title-case";
 export { baseFormConfig, createFormConfig } from "./config/hook-form-config";
 export { getCurrentUtcDate, getCurrentUtcISOString } from "./datetime";
 export { getNotificationColor, getNotificationIcon } from "./notification-ui-utils";
+export { createCalendarEventDetails } from "./create-calendar-event-details";
+export { createCalendarEvent, deleteCalendarEvent, getCalendarEvent, updateCalendarEvent } from "./google/calendar";
