+++ types/index.ts
@@ -27,3 +27,5 @@ export {
     type NotificationCardProps,
     notificationTypeGroups,
 } from "./actions/notification";
+export type { ActionResponse } from "./server-action-response";
+export type { Pets } from "./pets";
